pyproject.toml
[project]
name = "nitroplast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython>=1.84",
    # "diamond4py==0.0.3",
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "matplotlib-venn>=1.1.1",
    "matplotlib>=3.9.2",
    "pandas>=2.2.3",
    "tqdm>=4.67.0",
    "seaborn>=0.13.2",
    "pyopa>=0.8.4",
    "squarify>=0.4.4",
    "h5py>=3.15.1",
    "joblib>=1.5.3",
    "torch>=2.9.1",
    "scikit-learn>=1.8.0",
    "transformers>=4.57.3",
    "sentencepiece>=0.2.1",
    "umap-learn>=0.5.9.post2",
]
